feat(resolve) - Push column prop merging back to table
That's the right place for it. Now there's a hook for passing custom
`headerRows` per each header in case customization is needed (nested
case, to be documented).
