+++ packages/reactabular-resolve/README.md
@@ -2,6 +2,10 @@ Sometimes your rows might come in a nested format or it might have a representat
 
 ## API
 
+The API consists of two parts: data resolvers and column resolvers. If you have complex data, use the former. Latter come in handy if you have a nested column definition that needs to be flattened so that it works with Reactabular.
+
+## Data Resolvers
+
 Reactabular provides **resolve** module for handling with these cases. The system relies on an iterator that accepts rows and then transforms it using a resolver.
 
 ### `resolve.resolve`
@@ -30,11 +34,19 @@ The `byFunction` resolver accepts a path from where to look for a resolving func
 
 Instead of replacing the original value, `byFunction` generates `_<property>` kind of field to the resulting rows. Other functionality of Reactabular can use this hint and use the underscore field for user facing portions while using actual values for logic that relies on that.
 
+## Column Resolvers
+
 ### `resolve.columnChildren`
 
 **`({ columns, childrenField = 'children' }) => <resolved columns>`**
 
 Assuming your column definition is nested, you will need to resolve it to a flat format before other functionality of Reactabular can work. `resolve.columnChildren` does exactly that.
+
+### `resolve.headerRows`
+
+**`({ columns, childrenField = 'children' }) => <resolved columns>`**
+
+If your column definition is nested, you have to resolve it to header rows. `resolve.headerRows` has been designed exactly for this purpose.
 
 ## Combining Resolvers
