+++ packages/reactabular-resolve/package.json
@@ -33,9 +33,7 @@
   "main": "dist",
   "module": "dist",
   "jsnext:main": "dist",
-  "dependencies": {
-    "reactabular-utils": "^7.0.0"
-  },
+  "dependencies": {},
   "devDependencies": {
     "babel-cli": "^6.11.4",
     "rimraf": "^2.5.4"
