+++ packages/reactabular-table/src/header.js
@@ -1,5 +1,6 @@
 import React from 'react';
-import { tableHeaderContextTypes } from './types';
+import { mergePropPair } from 'reactabular-utils';
+import { tableHeaderTypes, tableHeaderContextTypes } from './types';
 import HeaderRow from './header-row';
 
 class Header extends React.Component { // eslint-disable-line max-len, react/prefer-stateless-function
@@ -9,17 +10,18 @@ class Header extends React.Component { // eslint-disable-line max-len, react/pre
     this.ref = null;
   }
   render() {
-    const { children, ...props } = this.props;
-    const { headerRows, components } = this.context;
+    const { children, headerRows, ...props } = this.props;
+    const { components, bodyColumns } = this.context;
 
     props.ref = (header) => {
       this.ref = header;
     };
 
+    // If headerRows aren't passed, default to bodyColumns as header rows
     return React.createElement(
       components.header.wrapper,
       props,
-      [headerRows.map((row, i) =>
+      [(mergeColumnProps(headerRows || [bodyColumns])).map((row, i) =>
         React.createElement(HeaderRow, {
           key: `${i}-header-row`,
           components: components.header,
@@ -32,9 +34,19 @@ class Header extends React.Component { // eslint-disable-line max-len, react/pre
     return this.ref;
   }
 }
-Header.propTypes = {
-  children: React.PropTypes.any
-};
+Header.propTypes = tableHeaderTypes;
 Header.contextTypes = tableHeaderContextTypes;
+
+function mergeColumnProps(columns) {
+  return columns.map(
+    row => row.map(column => (
+      column.header ? {
+        props: mergePropPair(column.props, column.header.props),
+        header: column.header,
+        column
+      } : {}
+    )
+  ));
+}
 
 export default Header;
